adds helper and implements try/catch so template won't error
