+++ ui/app/serializers/pki-certificate.js
@@ -2,7 +2,7 @@ import RESTSerializer from '@ember-data/serializer/rest';
 import { isNone, isBlank } from '@ember/utils';
 import { assign } from '@ember/polyfills';
 import { decamelize } from '@ember/string';
-import { pki } from 'node-forge';
+import { parsePkiCert } from '../helpers/parse-pki-cert';
 
 export default RESTSerializer.extend({
   keyForAttribute: function(attr) {
@@ -39,28 +39,34 @@ export default RESTSerializer.extend({
     return payload;
   },
 
-  getMetadata(model) {
-    // model is the responseJSON from the payload
-    const cert = pki.certificateFromPem(model.certificate);
-    const commonName = cert.subject.getField('CN').value;
-    const issueDate = cert.validity.notBefore;
-    const expiryDate = cert.validity.notAfter;
-    const certMetadata = {
-      common_name: commonName,
-      issue_date: issueDate,
-      expiry_date: expiryDate,
-    };
-    return certMetadata;
-  },
+  // getMetadata(model) {
+  //   // model is the responseJSON from the payload
+  //   const cert = pki.certificateFromPem(model.certificate);
+  //   const commonName = cert.subject.getField('CN').value;
+  //   const issueDate = cert.validity.notBefore;
+  //   const expiryDate = cert.validity.notAfter;
+  //   const certMetadata = {
+  //     common_name: commonName,
+  //     issue_date: issueDate,
+  //     expiry_date: expiryDate,
+  //   };
+  //   return certMetadata;
+  // },
 
   normalizeResponse(store, primaryModelClass, payload, id, requestType) {
     const responseJSON = this.normalizeItems(payload);
     const { modelName } = primaryModelClass;
-    const certMetadata = this.getMetadata(responseJSON);
+    let certMetadata;
+    try {
+      certMetadata = parsePkiCert([responseJSON]);
+    } catch {
+      certMetadata = null;
+    }
     // const certMetadata = getMetadata(responseJSON.certificate), return object with
     // getMetadata is a function (make a helper...later) use forge to parse the cert
     // get info we want, and always return an object so wrap in a try
     let transformedPayload = { [modelName]: { ...certMetadata, ...responseJSON } };
+    console.log(transformedPayload);
     return this._super(store, primaryModelClass, transformedPayload, id, requestType);
   },
